big refactor - all tests passing
