+++ planetmint/model/transaction.py
@@ -1,0 +1,2 @@
+
+class
