+++ tests/tendermint/test_core.py
@@ -17,7 +17,7 @@ from transactions.types.assets.create import Create
 from transactions.types.assets.transfer import Transfer
 from planetmint import App
 from planetmint.backend import query
-from planetmint.abci.core import OkCode, CodeTypeError, rollback
+from planetmint.abci.core import OkCode, CodeTypeError
 from planetmint.abci.block import Block
 from planetmint.abci.tendermint_utils import new_validator_set
 from planetmint.abci.tendermint_utils import public_key_to_base64
@@ -51,9 +51,9 @@ def test_init_chain_successfully_registers_chain(b):
     request = generate_init_chain_request("chain-XYZ")
     res = App(b).init_chain(request)
     assert res == types.ResponseInitChain()
-    chain = query.get_latest_abci_chain(b.connection)
+    chain = query.get_latest_abci_chain(b.models.connection)
     assert chain == {"height": 0, "chain_id": "chain-XYZ", "is_synced": True}
-    assert query.get_latest_block(b.connection) == {
+    assert query.get_latest_block(b.models.connection) == {
         "height": 0,
         "app_hash": "",
         "transaction_ids": [],
@@ -66,7 +66,7 @@ def test_init_chain_ignores_invalid_init_chain_requests(b):
     res = App(b).init_chain(request)
     assert res == types.ResponseInitChain()
 
-    validator_set = query.get_validator_set(b.connection)
+    validator_set = query.get_validator_set(b.models.connection)
 
     invalid_requests = [
         request,  # the same request again
@@ -79,11 +79,11 @@ def test_init_chain_ignores_invalid_init_chain_requests(b):
         with pytest.raises(SystemExit):
             App(b).init_chain(r)
         # assert nothing changed - neither validator set, nor chain ID
-        new_validator_set = query.get_validator_set(b.connection)
+        new_validator_set = query.get_validator_set(b.models.connection)
         assert new_validator_set == validator_set
-        new_chain_id = query.get_latest_abci_chain(b.connection)["chain_id"]
+        new_chain_id = query.get_latest_abci_chain(b.models.connection)["chain_id"]
         assert new_chain_id == "chain-XYZ"
-        assert query.get_latest_block(b.connection) == {
+        assert query.get_latest_block(b.models.connection) == {
             "height": 0,
             "app_hash": "",
             "transaction_ids": [],
@@ -96,10 +96,10 @@ def test_init_chain_recognizes_new_chain_after_migration(b):
     res = App(b).init_chain(request)
     assert res == types.ResponseInitChain()
 
-    validator_set = query.get_validator_set(b.connection)["validators"]
+    validator_set = query.get_validator_set(b.models.connection)["validators"]
 
     # simulate a migration
-    query.store_block(b.connection, Block(app_hash="", height=1, transactions=[])._asdict())
+    query.store_block(b.models.connection, Block(app_hash="", height=1, transactions=[])._asdict())
     b.migrate_abci_chain()
 
     # the same or other mismatching requests are ignored
@@ -112,12 +112,12 @@ def test_init_chain_recognizes_new_chain_after_migration(b):
     for r in invalid_requests:
         with pytest.raises(SystemExit):
             App(b).init_chain(r)
-        assert query.get_latest_abci_chain(b.connection) == {
+        assert query.get_latest_abci_chain(b.models.connection) == {
             "chain_id": "chain-XYZ-migrated-at-height-1",
             "is_synced": False,
             "height": 2,
         }
-        new_validator_set = query.get_validator_set(b.connection)["validators"]
+        new_validator_set = query.get_validator_set(b.models.connection)["validators"]
         assert new_validator_set == validator_set
 
     # a request with the matching chain ID and matching validator set
@@ -125,12 +125,12 @@ def test_init_chain_recognizes_new_chain_after_migration(b):
     request = generate_init_chain_request("chain-XYZ-migrated-at-height-1", validators)
     res = App(b).init_chain(request)
     assert res == types.ResponseInitChain()
-    assert query.get_latest_abci_chain(b.connection) == {
+    assert query.get_latest_abci_chain(b.models.connection) == {
         "chain_id": "chain-XYZ-migrated-at-height-1",
         "is_synced": True,
         "height": 2,
     }
-    assert query.get_latest_block(b.connection) == {
+    assert query.get_latest_block(b.models.connection) == {
         "height": 2,
         "app_hash": "",
         "transaction_ids": [],
@@ -145,14 +145,14 @@ def test_init_chain_recognizes_new_chain_after_migration(b):
     for r in invalid_requests:
         with pytest.raises(SystemExit):
             App(b).init_chain(r)
-        assert query.get_latest_abci_chain(b.connection) == {
+        assert query.get_latest_abci_chain(b.models.connection) == {
             "chain_id": "chain-XYZ-migrated-at-height-1",
             "is_synced": True,
             "height": 2,
         }
-        new_validator_set = query.get_validator_set(b.connection)["validators"]
+        new_validator_set = query.get_validator_set(b.models.connection)["validators"]
         assert new_validator_set == validator_set
-        assert query.get_latest_block(b.connection) == {
+        assert query.get_latest_block(b.models.connection) == {
             "height": 2,
             "app_hash": "",
             "transaction_ids": [],
@@ -167,28 +167,28 @@ def test_info(b):
     assert res.last_block_height == 0
     assert res.last_block_app_hash == b""
 
-    b.store_block(Block(app_hash="1", height=1, transactions=[])._asdict())
+    b.models.store_block(Block(app_hash="1", height=1, transactions=[])._asdict())
     res = app.info(r)
     assert res.last_block_height == 1
     assert res.last_block_app_hash == b"1"
 
     # simulate a migration and assert the height is shifted
-    b.store_abci_chain(2, "chain-XYZ")
+    b.models.store_abci_chain(2, "chain-XYZ")
     app = App(b)
-    b.store_block(Block(app_hash="2", height=2, transactions=[])._asdict())
+    b.models.store_block(Block(app_hash="2", height=2, transactions=[])._asdict())
     res = app.info(r)
     assert res.last_block_height == 0
     assert res.last_block_app_hash == b"2"
 
-    b.store_block(Block(app_hash="3", height=3, transactions=[])._asdict())
+    b.models.store_block(Block(app_hash="3", height=3, transactions=[])._asdict())
     res = app.info(r)
     assert res.last_block_height == 1
     assert res.last_block_app_hash == b"3"
 
     # it's always the latest migration that is taken into account
-    b.store_abci_chain(4, "chain-XYZ-new")
+    b.models.store_abci_chain(4, "chain-XYZ-new")
     app = App(b)
-    b.store_block(Block(app_hash="4", height=4, transactions=[])._asdict())
+    b.models.store_block(Block(app_hash="4", height=4, transactions=[])._asdict())
     res = app.info(r)
     assert res.last_block_height == 0
     assert res.last_block_app_hash == b"4"
@@ -235,7 +235,7 @@ def test_deliver_tx__valid_create_updates_db_and_emits_event(b, init_chain_reque
 
     app.end_block(types.RequestEndBlock(height=99))
     app.commit()
-    assert b.get_transaction(tx.id).id == tx.id
+    assert b.models.get_transaction(tx.id).id == tx.id
     block_event = events.get()
     assert block_event.data["transactions"] == [tx]
 
@@ -264,7 +264,7 @@ def test_deliver_tx__double_spend_fails(b, init_chain_request):
 
     app.end_block(types.RequestEndBlock(height=99))
     app.commit()
-    assert b.get_transaction(tx.id).id == tx.id
+    assert b.models.get_transaction(tx.id).id == tx.id
     result = app.deliver_tx(encode_tx_to_bytes(tx))
     assert result.code == CodeTypeError
 
@@ -311,7 +311,7 @@ def test_end_block_return_validator_updates(b, init_chain_request):
 
     # generate a block containing a concluded validator election
     validators = generate_validators([1] * 4)
-    b.store_validator_set(1, [v["storage"] for v in validators])
+    b.models.store_validator_set(1, [v["storage"] for v in validators])
 
     new_validator = generate_validators([1])[0]
 
@@ -322,8 +322,8 @@ def test_end_block_return_validator_updates(b, init_chain_request):
     election, votes = generate_election(
         b, ValidatorElection, public_key, private_key, [{"data": new_validator["election"]}], voter_keys
     )
-    b.store_block(Block(height=1, transactions=[election.id], app_hash="")._asdict())
-    b.store_bulk_transactions([election])
+    b.models.store_block(Block(height=1, transactions=[election.id], app_hash="")._asdict())
+    b.models.store_bulk_transactions([election])
     b.process_block(1, [election])
 
     app.block_transactions = votes
@@ -352,32 +352,32 @@ def test_store_pre_commit_state_in_end_block(b, alice, init_chain_request):
     app.deliver_tx(encode_tx_to_bytes(tx))
     app.end_block(types.RequestEndBlock(height=99))
 
-    resp = query.get_pre_commit_state(b.connection)
+    resp = query.get_pre_commit_state(b.models.connection)
     assert resp["height"] == 99
     assert resp["transactions"] == [tx.id]
 
     app.begin_block(begin_block)
     app.deliver_tx(encode_tx_to_bytes(tx))
     app.end_block(types.RequestEndBlock(height=100))
-    resp = query.get_pre_commit_state(b.connection)
+    resp = query.get_pre_commit_state(b.models.connection)
     assert resp["height"] == 100
     assert resp["transactions"] == [tx.id]
 
     # simulate a chain migration and assert the height is shifted
-    b.store_abci_chain(100, "new-chain")
+    b.models.store_abci_chain(100, "new-chain")
     app = App(b)
     app.begin_block(begin_block)
     app.deliver_tx(encode_tx_to_bytes(tx))
     app.end_block(types.RequestEndBlock(height=1))
-    resp = query.get_pre_commit_state(b.connection)
+    resp = query.get_pre_commit_state(b.models.connection)
     assert resp["height"] == 101
     assert resp["transactions"] == [tx.id]
 
 
-def test_rollback_pre_commit_state_after_crash(b):
+def test_rollback_pre_commit_state_after_crash(b, test_models):
     validators = generate_validators([1] * 4)
-    b.store_validator_set(1, [v["storage"] for v in validators])
-    b.store_block(Block(height=1, transactions=[], app_hash="")._asdict())
+    b.models.store_validator_set(1, [v["storage"] for v in validators])
+    b.models.store_block(Block(height=1, transactions=[], app_hash="")._asdict())
 
     public_key = validators[0]["public_key"]
     private_key = validators[0]["private_key"]
@@ -398,35 +398,35 @@ def test_rollback_pre_commit_state_after_crash(b):
     total_votes += votes
     txs += [validator_election, *votes]
 
-    b.store_bulk_transactions(txs)
-    b.store_abci_chain(2, "new_chain")
-    b.store_validator_set(2, [v["storage"] for v in validators])
+    b.models.store_bulk_transactions( txs)
+    b.models.store_abci_chain(2, "new_chain")
+    b.models.store_validator_set(2, [v["storage"] for v in validators])
     # TODO change to `4` when upgrading to Tendermint 0.22.4.
-    b.store_validator_set(3, [new_validator["storage"]])
-    b.store_election(migration_election.id, 2, is_concluded=False)
-    b.store_election(validator_election.id, 2, is_concluded=True)
+    b.models.store_validator_set(3, [new_validator["storage"]])
+    b.models.store_election(migration_election.id, 2, is_concluded=False)
+    b.models.store_election(validator_election.id, 2, is_concluded=True)
 
     # no pre-commit state
-    rollback(b)
+    b.rollback()
 
     for tx in txs:
-        assert b.get_transaction(tx.id)
-    assert b.get_latest_abci_chain()
-    assert len(b.get_validator_set()["validators"]) == 1
-    assert b.get_election(migration_election.id)
-    assert b.get_election(validator_election.id)
+        assert b.models.get_transaction(tx.id)
+    assert b.models.get_latest_abci_chain()
+    assert len(b.models.get_validator_set()["validators"]) == 1
+    assert b.models.get_election(migration_election.id)
+    assert b.models.get_election(validator_election.id)
 
-    b.store_pre_commit_state({"height": 2, "transactions": [tx.id for tx in txs]})
+    b.models.store_pre_commit_state({"height": 2, "transactions": [tx.id for tx in txs]})
 
-    rollback(b)
+    b.rollback()
 
     for tx in txs:
-        assert not b.get_transaction(tx.id)
-    assert not b.get_latest_abci_chain()
-    assert len(b.get_validator_set()["validators"]) == 4
-    assert len(b.get_validator_set(2)["validators"]) == 4
-    assert not b.get_election(migration_election.id)
-    assert not b.get_election(validator_election.id)
+        assert not b.models.get_transaction(tx.id)
+    assert not b.models.get_latest_abci_chain()
+    assert len(b.models.get_validator_set()["validators"]) == 4
+    assert len(b.models.get_validator_set(2)["validators"]) == 4
+    assert not b.models.get_election(migration_election.id)
+    assert not b.models.get_election(validator_election.id)
 
 
 def test_new_validator_set(b):
@@ -451,8 +451,8 @@ def test_new_validator_set(b):
 
     validators = [node1]
     updates = [node1_new_power, node2]
-    b.store_validator_set(1, validators)
-    updated_validator_set = new_validator_set(b.get_validators(1), updates)
+    b.models.store_validator_set(1, validators)
+    updated_validator_set = new_validator_set(b.models.get_validators(1), updates)
 
     updated_validators = []
     for u in updates:
@@ -467,42 +467,42 @@ def test_new_validator_set(b):
 
 
 def test_info_aborts_if_chain_is_not_synced(b):
-    b.store_abci_chain(0, "chain-XYZ", False)
+    b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
         App(b).info(types.RequestInfo())
 
 
 def test_check_tx_aborts_if_chain_is_not_synced(b):
-    b.store_abci_chain(0, "chain-XYZ", False)
+    b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
         App(b).check_tx("some bytes")
 
 
 def test_begin_aborts_if_chain_is_not_synced(b):
-    b.store_abci_chain(0, "chain-XYZ", False)
+    b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
         App(b).info(types.RequestBeginBlock())
 
 
 def test_deliver_tx_aborts_if_chain_is_not_synced(b):
-    b.store_abci_chain(0, "chain-XYZ", False)
+    b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
         App(b).deliver_tx("some bytes")
 
 
 def test_end_block_aborts_if_chain_is_not_synced(b):
-    b.store_abci_chain(0, "chain-XYZ", False)
+    b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
         App(b).info(types.RequestEndBlock())
 
 
 def test_commit_aborts_if_chain_is_not_synced(b):
-    b.store_abci_chain(0, "chain-XYZ", False)
+    b.models.store_abci_chain(0, "chain-XYZ", False)
 
     with pytest.raises(SystemExit):
         App(b).commit()
