+++ tests/tendermint/test_integration.py
@@ -34,16 +34,16 @@ def test_app(b, eventqueue_fixture, init_chain_request):
     assert res
     assert res.info.last_block_app_hash == b""
     assert res.info.last_block_height == 0
-    assert not b.get_latest_block()
+    assert not b.models.get_latest_block()
 
     p.process("init_chain", types.Request(init_chain=init_chain_request))
-    block0 = b.get_latest_block()
+    block0 = b.models.get_latest_block()
     assert block0
     assert block0["height"] == 0
     assert block0["app_hash"] == ""
 
     pk = codecs.encode(init_chain_request.validators[0].pub_key.ed25519, "base64").decode().strip("\n")
-    [validator] = b.get_validators(height=1)
+    [validator] = b.models.get_validators(height=1)
     assert validator["public_key"]["value"] == pk
     assert validator["voting_power"] == 10
 
@@ -81,9 +81,9 @@ def test_app(b, eventqueue_fixture, init_chain_request):
     data = p.process("commit", None)
     res = next(read_messages(BytesIO(data), types.Response))
     assert res.commit.data == new_block_hash.encode("utf-8")
-    assert b.get_transaction(tx.id).id == tx.id
+    assert b.models.get_transaction(tx.id).id == tx.id
 
-    block0 = b.get_latest_block()
+    block0 = b.models.get_latest_block()
     assert block0
     assert block0["height"] == 1
     assert block0["app_hash"] == new_block_hash
@@ -101,7 +101,7 @@ def test_app(b, eventqueue_fixture, init_chain_request):
     res = next(read_messages(BytesIO(data), types.Response))
     assert res.commit.data == new_block_hash.encode("utf-8")
 
-    block0 = b.get_latest_block()
+    block0 = b.models.get_latest_block()
     assert block0
     assert block0["height"] == 2
 
@@ -110,7 +110,7 @@ def test_app(b, eventqueue_fixture, init_chain_request):
 
 
 @pytest.mark.abci
-def test_post_transaction_responses(tendermint_ws_url, b):
+def test_post_transaction_responses(tendermint_ws_url, b, test_abci_rpc):
     from transactions.common.crypto import generate_key_pair
 
     alice = generate_key_pair()
@@ -121,14 +121,18 @@ def test_post_transaction_responses(tendermint_ws_url, b):
         assets=[{"data": "QmaozNR7DZHQK1ZcU9p7QdrshMvXqWK6gpu5rmrkPdT3L4"}],
     ).sign([alice.private_key])
 
-    code, message = ABCI_RPC().write_transaction(MODE_LIST, b.tendermint_rpc_endpoint, MODE_COMMIT, tx, BROADCAST_TX_COMMIT)
+    code, message = test_abci_rpc.write_transaction(
+        MODE_LIST, test_abci_rpc.tendermint_rpc_endpoint, MODE_COMMIT, tx, BROADCAST_TX_COMMIT
+    )
     assert code == 202
 
     tx_transfer = Transfer.generate(tx.to_inputs(), [([bob.public_key], 1)], asset_ids=[tx.id]).sign(
         [alice.private_key]
     )
 
-    code, message = ABCI_RPC().write_transaction(MODE_LIST, b.tendermint_rpc_endpoint, MODE_COMMIT, tx_transfer, BROADCAST_TX_COMMIT)
+    code, message = test_abci_rpc.write_transaction(
+        MODE_LIST, test_abci_rpc.tendermint_rpc_endpoint, MODE_COMMIT, tx_transfer, BROADCAST_TX_COMMIT
+    )
     assert code == 202
 
     carly = generate_key_pair()
@@ -138,6 +142,8 @@ def test_post_transaction_responses(tendermint_ws_url, b):
         asset_ids=[tx.id],
     ).sign([alice.private_key])
     for mode in (BROADCAST_TX_SYNC, BROADCAST_TX_COMMIT):
-        code, message = ABCI_RPC().write_transaction(MODE_LIST, b.tendermint_rpc_endpoint, MODE_COMMIT, double_spend, mode)
+        code, message = test_abci_rpc.write_transaction(
+            MODE_LIST, test_abci_rpc.tendermint_rpc_endpoint, MODE_COMMIT, double_spend, mode
+        )
         assert code == 500
         assert message == "Transaction validation failed"
