+++ tests/tendermint/test_lib.py
@@ -24,14 +24,13 @@ from uuid import uuid4
 from planetmint.abci.rpc import ABCI_RPC
 from planetmint.abci.rpc import MODE_COMMIT, MODE_LIST
 
-
 @pytest.mark.bdb
 def test_asset_is_separated_from_transaciton(b):
     import copy
     from transactions.common.crypto import generate_key_pair
     from planetmint.backend.tarantool.connection import TarantoolDBConnection
 
-    if isinstance(b.connection, TarantoolDBConnection):
+    if isinstance(b.models.connection, TarantoolDBConnection):
         pytest.skip("This specific function is skipped because, assets are stored differently if using Tarantool")
 
     alice = generate_key_pair()
@@ -61,16 +60,16 @@ def test_asset_is_separated_from_transaciton(b):
         [alice.private_key]
     )
 
-    # with store_bulk_transactions we use `insert_many` where PyMongo
+    # with b.models.store_bulk_transactions we use `insert_many` where PyMongo
     # automatically adds an `_id` field to the tx, therefore we need the
     # deepcopy, for more info see:
     # https://api.mongodb.com/python/current/faq.html#writes-and-ids
     tx_dict = copy.deepcopy(tx.to_dict())
 
-    b.store_bulk_transactions([tx])
-    assert "asset" not in backend.query.get_transaction_single(b.connection, tx.id)
-    assert backend.query.get_asset(b.connection, tx.id).data == assets[0]
-    assert b.get_transaction(tx.id).to_dict() == tx_dict
+    b.models.store_bulk_transactions( [tx])
+    assert "asset" not in backend.query.get_transaction_single(b.models.connection, tx.id)
+    assert backend.query.get_asset(b.models.connection, tx.id).data == assets[0]
+    assert b.models.get_transaction(tx.id).to_dict() == tx_dict
 
 
 @pytest.mark.bdb
@@ -81,9 +80,9 @@ def test_get_latest_block(b):
         app_hash = os.urandom(16).hex()
         txn_id = os.urandom(16).hex()
         block = Block(app_hash=app_hash, height=i, transactions=[txn_id])._asdict()
-        b.store_block(block)
+        b.models.store_block(block)
 
-    block = b.get_latest_block()
+    block = b.models.get_latest_block()
     assert block["height"] == 9
 
 
@@ -100,7 +99,7 @@ def test_validation_error(b):
 
 
 @patch("requests.post")
-def test_write_and_post_transaction(mock_post, b):
+def test_write_and_post_transaction(mock_post, b, test_abci_rpc):
     from transactions.common.crypto import generate_key_pair
     from planetmint.abci.tendermint_utils import encode_transaction
 
@@ -110,7 +109,7 @@ def test_write_and_post_transaction(mock_post, b):
     )
 
     tx = b.validate_transaction(tx)
-    ABCI_RPC().write_transaction(MODE_LIST, b.tendermint_rpc_endpoint, MODE_COMMIT, tx, BROADCAST_TX_ASYNC)
+    test_abci_rpc.write_transaction(MODE_LIST, test_abci_rpc.tendermint_rpc_endpoint, MODE_COMMIT, tx, BROADCAST_TX_ASYNC)
 
     assert mock_post.called
     args, kwargs = mock_post.call_args
@@ -121,7 +120,7 @@ def test_write_and_post_transaction(mock_post, b):
 
 @patch("requests.post")
 @pytest.mark.parametrize("mode", [BROADCAST_TX_SYNC, BROADCAST_TX_ASYNC, BROADCAST_TX_COMMIT])
-def test_post_transaction_valid_modes(mock_post, b, mode):
+def test_post_transaction_valid_modes(mock_post, b, mode, test_abci_rpc):
     from transactions.common.crypto import generate_key_pair
 
     alice = generate_key_pair()
@@ -129,13 +128,13 @@ def test_post_transaction_valid_modes(mock_post, b, mode):
         Create.generate([alice.public_key], [([alice.public_key], 1)], assets=None).sign([alice.private_key]).to_dict()
     )
     tx = b.validate_transaction(tx)
-    ABCI_RPC().write_transaction(MODE_LIST, b.tendermint_rpc_endpoint, MODE_COMMIT, tx, mode)
+    test_abci_rpc.write_transaction(MODE_LIST, test_abci_rpc.tendermint_rpc_endpoint, MODE_COMMIT, tx, mode)
 
     args, kwargs = mock_post.call_args
     assert mode == kwargs["json"]["method"]
 
 
-def test_post_transaction_invalid_mode(b):
+def test_post_transaction_invalid_mode(b, test_abci_rpc):
     from transactions.common.crypto import generate_key_pair
     from transactions.common.exceptions import ValidationError
 
@@ -145,18 +144,18 @@ def test_post_transaction_invalid_mode(b):
     )
     tx = b.validate_transaction(tx)
     with pytest.raises(ValidationError):
-        ABCI_RPC().write_transaction(MODE_LIST, b.tendermint_rpc_endpoint, MODE_COMMIT, tx, "nope")
+        test_abci_rpc.write_transaction(MODE_LIST, test_abci_rpc.tendermint_rpc_endpoint, MODE_COMMIT, tx, "nope")
 
 
 @pytest.mark.bdb
 def test_update_utxoset(b, signed_create_tx, signed_transfer_tx, db_conn):
-    b.update_utxoset(signed_create_tx)
+    b.models.tests_update_utxoset(signed_create_tx)
     utxoset = db_conn.get_space("utxos")
     assert utxoset.select().rowcount == 1
     utxo = utxoset.select().data
     assert utxo[0][1] == signed_create_tx.id
     assert utxo[0][2] == 0
-    b.update_utxoset(signed_transfer_tx)
+    b.models.tests_update_utxoset(signed_transfer_tx)
     assert utxoset.select().rowcount == 1
     utxo = utxoset.select().data
     assert utxo[0][1] == signed_transfer_tx.id
@@ -166,26 +165,26 @@ def test_update_utxoset(b, signed_create_tx, signed_transfer_tx, db_conn):
 @pytest.mark.bdb
 def test_store_transaction(mocker, b, signed_create_tx, signed_transfer_tx):
     mocked_store_transaction = mocker.patch("planetmint.backend.query.store_transactions")
-    b.store_bulk_transactions([signed_create_tx])
-    mocked_store_transaction.assert_any_call(b.connection, [signed_create_tx.to_dict()], "transactions")
+    b.models.store_bulk_transactions( [signed_create_tx])
+    mocked_store_transaction.assert_any_call(b.models.connection, [signed_create_tx.to_dict()], "transactions")
     mocked_store_transaction.reset_mock()
-    b.store_bulk_transactions([signed_transfer_tx])
+    b.models.store_bulk_transactions( [signed_transfer_tx])
 
 
 @pytest.mark.bdb
 def test_store_bulk_transaction(mocker, b, signed_create_tx, signed_transfer_tx):
     mocked_store_transactions = mocker.patch("planetmint.backend.query.store_transactions")
-    b.store_bulk_transactions((signed_create_tx,))
-    mocked_store_transactions.assert_any_call(b.connection, [signed_create_tx.to_dict()], "transactions")
+    b.models.store_bulk_transactions( (signed_create_tx,))
+    mocked_store_transactions.assert_any_call(b.models.connection, [signed_create_tx.to_dict()], "transactions")
     mocked_store_transactions.reset_mock()
-    b.store_bulk_transactions((signed_transfer_tx,))
+    b.models.store_bulk_transactions( (signed_transfer_tx,))
 
 
 @pytest.mark.bdb
 def test_delete_zero_unspent_outputs(b, utxoset):
     unspent_outputs, utxo_collection = utxoset
     num_rows_before_operation = utxo_collection.select().rowcount
-    delete_res = b.delete_unspent_outputs()  # noqa: F841
+    delete_res = b.models.tests_delete_unspent_outputs()  # noqa: F841
     num_rows_after_operation = utxo_collection.select().rowcount
     # assert delete_res is None
     assert num_rows_before_operation == num_rows_after_operation
@@ -193,12 +192,12 @@ def test_delete_zero_unspent_outputs(b, utxoset):
 
 @pytest.mark.bdb
 def test_delete_one_unspent_outputs(b, dummy_unspent_outputs):
-    utxo_space = b.connection.get_space("utxos")
+    utxo_space = b.models.connection.get_space("utxos")
     for utxo in dummy_unspent_outputs:
         res = utxo_space.insert((uuid4().hex, utxo["transaction_id"], utxo["output_index"], utxo))
         assert res
 
-    b.delete_unspent_outputs(dummy_unspent_outputs[0])
+    b.models.tests_delete_unspent_outputs(dummy_unspent_outputs[0])
     res1 = utxo_space.select(["a", 1], index="utxo_by_transaction_id_and_output_index").data
     res2 = utxo_space.select(["b", 0], index="utxo_by_transaction_id_and_output_index").data
     assert len(res1) + len(res2) == 2
@@ -208,12 +207,12 @@ def test_delete_one_unspent_outputs(b, dummy_unspent_outputs):
 
 @pytest.mark.bdb
 def test_delete_many_unspent_outputs(b, dummy_unspent_outputs):
-    utxo_space = b.connection.get_space("utxos")
+    utxo_space = b.models.connection.get_space("utxos")
     for utxo in dummy_unspent_outputs:
         res = utxo_space.insert((uuid4().hex, utxo["transaction_id"], utxo["output_index"], utxo))
         assert res
 
-    b.delete_unspent_outputs(*dummy_unspent_outputs[::2])
+    b.models.tests_delete_unspent_outputs(*dummy_unspent_outputs[::2])
     res1 = utxo_space.select(["a", 0], index="utxo_by_transaction_id_and_output_index").data
     res2 = utxo_space.select(["b", 0], index="utxo_by_transaction_id_and_output_index").data
     assert len(res1) + len(res2) == 0
@@ -223,9 +222,9 @@ def test_delete_many_unspent_outputs(b, dummy_unspent_outputs):
 
 @pytest.mark.bdb
 def test_store_zero_unspent_output(b):
-    utxos = b.connection.get_space("utxos")
+    utxos = b.models.connection.get_space("utxos")
     num_rows_before_operation = utxos.select().rowcount
-    res = b.store_unspent_outputs()
+    res = b.models.tests_store_unspent_outputs()
     num_rows_after_operation = utxos.select().rowcount
     assert res is None
     assert num_rows_before_operation == num_rows_after_operation
@@ -235,8 +234,8 @@ def test_store_zero_unspent_output(b):
 def test_store_one_unspent_output(b, unspent_output_1, utxo_collection):
     from planetmint.backend.tarantool.connection import TarantoolDBConnection
 
-    res = b.store_unspent_outputs(unspent_output_1)
-    if not isinstance(b.connection, TarantoolDBConnection):
+    res = b.models.tests_store_unspent_outputs(unspent_output_1)
+    if not isinstance(b.models.connection, TarantoolDBConnection):
         assert res.acknowledged
         assert len(list(res)) == 1
         assert (
@@ -249,7 +248,7 @@ def test_store_one_unspent_output(b, unspent_output_1, utxo_collection):
             == 1
         )
     else:
-        utx_space = b.connection.get_space("utxos")
+        utx_space = b.models.connection.get_space("utxos")
         res = utx_space.select(
             [unspent_output_1["transaction_id"], unspent_output_1["output_index"]],
             index="utxo_by_transaction_id_and_output_index",
@@ -259,25 +258,25 @@ def test_store_one_unspent_output(b, unspent_output_1, utxo_collection):
 
 @pytest.mark.bdb
 def test_store_many_unspent_outputs(b, unspent_outputs):
-    b.store_unspent_outputs(*unspent_outputs)
-    utxo_space = b.connection.get_space("utxos")
+    b.models.tests_store_unspent_outputs(*unspent_outputs)
+    utxo_space = b.models.connection.get_space("utxos")
     res = utxo_space.select([unspent_outputs[0]["transaction_id"]], index="utxos_by_transaction_id")
     assert len(res.data) == 3
 
 
 def test_get_utxoset_merkle_root_when_no_utxo(b):
-    assert b.get_utxoset_merkle_root() == sha3_256(b"").hexdigest()
+    assert b.models.tests_get_utxoset_merkle_root() == sha3_256(b"").hexdigest()
 
 
 @pytest.mark.bdb
 def test_get_utxoset_merkle_root(b, dummy_unspent_outputs):
-    utxo_space = b.connection.get_space("utxos")
+    utxo_space = b.models.connection.get_space("utxos")
     for utxo in dummy_unspent_outputs:
         res = utxo_space.insert((uuid4().hex, utxo["transaction_id"], utxo["output_index"], utxo))
         assert res
 
     expected_merkle_root = "86d311c03115bf4d287f8449ca5828505432d69b82762d47077b1c00fe426eac"
-    merkle_root = b.get_utxoset_merkle_root()
+    merkle_root = b.models.tests_get_utxoset_merkle_root()
     assert merkle_root == expected_merkle_root
 
 
@@ -301,20 +300,20 @@ def test_get_spent_transaction_double_spend(b, alice, bob, carol):
         tx.to_inputs() + tx.to_inputs(), [([bob.public_key], 1)], asset_ids=[tx.id]
     ).sign([alice.private_key])
 
-    b.store_bulk_transactions([tx])
+    b.models.store_bulk_transactions( [tx])
 
     with pytest.raises(DoubleSpend):
         b.validate_transaction(same_input_double_spend)
 
-    assert b.get_spent(tx.id, tx_transfer.inputs[0].fulfills.output, [tx_transfer])
+    assert b.models.get_spent(tx.id, tx_transfer.inputs[0].fulfills.output, [tx_transfer])
 
     with pytest.raises(DoubleSpend):
-        b.get_spent(tx.id, tx_transfer.inputs[0].fulfills.output, [tx_transfer, double_spend])
+        b.models.get_spent(tx.id, tx_transfer.inputs[0].fulfills.output, [tx_transfer, double_spend])
 
-    b.store_bulk_transactions([tx_transfer])
+    b.models.store_bulk_transactions( [tx_transfer])
 
     with pytest.raises(DoubleSpend):
-        b.get_spent(tx.id, tx_transfer.inputs[0].fulfills.output, [double_spend])
+        b.models.get_spent(tx.id, tx_transfer.inputs[0].fulfills.output, [double_spend])
 
 
 def test_validation_with_transaction_buffer(b):
@@ -341,7 +340,7 @@ def test_validation_with_transaction_buffer(b):
 @pytest.mark.bdb
 def test_migrate_abci_chain_yields_on_genesis(b):
     b.migrate_abci_chain()
-    latest_chain = b.get_latest_abci_chain()
+    latest_chain = b.models.get_latest_abci_chain()
     assert latest_chain is None
 
 
@@ -362,10 +361,10 @@ def test_migrate_abci_chain_yields_on_genesis(b):
     ],
 )
 def test_migrate_abci_chain_generates_new_chains(b, chain, block_height, expected):
-    b.store_abci_chain(*chain)
-    b.store_block(Block(app_hash="", height=block_height, transactions=[])._asdict())
+    b.models.store_abci_chain(*chain)
+    b.models.store_block(Block(app_hash="", height=block_height, transactions=[])._asdict())
     b.migrate_abci_chain()
-    latest_chain = b.get_latest_abci_chain()
+    latest_chain = b.models.get_latest_abci_chain()
     assert latest_chain == expected
 
 
@@ -378,13 +377,13 @@ def test_get_spent_key_order(b, user_pk, user_sk, user2_pk, user2_sk):
     bob = generate_key_pair()
 
     tx1 = Create.generate([user_pk], [([alice.public_key], 3), ([user_pk], 2)]).sign([user_sk])
-    b.store_bulk_transactions([tx1])
+    b.models.store_bulk_transactions( [tx1])
 
     inputs = tx1.to_inputs()
     tx2 = Transfer.generate([inputs[1]], [([user2_pk], 2)], [tx1.id]).sign([user_sk])
     assert b.validate_transaction(tx2)
 
-    b.store_bulk_transactions([tx2])
+    b.models.store_bulk_transactions( [tx2])
 
     tx3 = Transfer.generate([inputs[1]], [([bob.public_key], 2)], [tx1.id]).sign([user_sk])
