+++ tests/test_core.py
@@ -5,7 +5,6 @@
 
 import pytest
 
-from planetmint.version import __tm_supported_versions__
 from transactions.types.assets.create import Create
 from transactions.types.assets.transfer import Transfer
 
@@ -39,13 +38,13 @@ def config(request, monkeypatch):
 
 
 def test_bigchain_class_default_initialization(config):
-    from planetmint import Planetmint
-    from planetmint.validation import BaseValidationRules
+    from planetmint.application.validation import Validator
+    from planetmint.application.basevalidationrules import BaseValidationRules
 
-    planet = Planetmint()
-    assert planet.connection.host == config["database"]["host"]
-    assert planet.connection.port == config["database"]["port"]
-    assert planet.validation == BaseValidationRules
+    validator = Validator()
+    assert validator.models.connection.host == config["database"]["host"]
+    assert validator.models.connection.port == config["database"]["port"]
+    assert validator.validation == BaseValidationRules
 
 
 @pytest.mark.bdb
@@ -55,7 +54,7 @@ def test_get_spent_issue_1271(b, alice, bob, carol):
         [([carol.public_key], 8)],
     ).sign([carol.private_key])
     assert b.validate_transaction(tx_1)
-    b.store_bulk_transactions([tx_1])
+    b.models.store_bulk_transactions( [tx_1])
 
     tx_2 = Transfer.generate(
         tx_1.to_inputs(),
@@ -63,7 +62,7 @@ def test_get_spent_issue_1271(b, alice, bob, carol):
         asset_ids=[tx_1.id],
     ).sign([carol.private_key])
     assert b.validate_transaction(tx_2)
-    b.store_bulk_transactions([tx_2])
+    b.models.store_bulk_transactions( [tx_2])
 
     tx_3 = Transfer.generate(
         tx_2.to_inputs()[2:3],
@@ -71,7 +70,7 @@ def test_get_spent_issue_1271(b, alice, bob, carol):
         asset_ids=[tx_1.id],
     ).sign([carol.private_key])
     assert b.validate_transaction(tx_3)
-    b.store_bulk_transactions([tx_3])
+    b.models.store_bulk_transactions( [tx_3])
 
     tx_4 = Transfer.generate(
         tx_2.to_inputs()[1:2] + tx_3.to_inputs()[0:1],
@@ -79,7 +78,7 @@ def test_get_spent_issue_1271(b, alice, bob, carol):
         asset_ids=[tx_1.id],
     ).sign([alice.private_key])
     assert b.validate_transaction(tx_4)
-    b.store_bulk_transactions([tx_4])
+    b.models.store_bulk_transactions( [tx_4])
 
     tx_5 = Transfer.generate(
         tx_2.to_inputs()[0:1],
@@ -88,8 +87,8 @@ def test_get_spent_issue_1271(b, alice, bob, carol):
     ).sign([bob.private_key])
     assert b.validate_transaction(tx_5)
 
-    b.store_bulk_transactions([tx_5])
-    assert b.get_spent(tx_2.id, 0) == tx_5.to_dict()
-    assert not b.get_spent(tx_5.id, 0)
-    assert b.get_outputs_filtered(alice.public_key)
-    assert b.get_outputs_filtered(alice.public_key, spent=False)
+    b.models.store_bulk_transactions( [tx_5])
+    assert b.models.get_spent(tx_2.id, 0) == tx_5.to_dict()
+    assert not b.models.get_spent(tx_5.id, 0)
+    assert b.models.get_outputs_filtered(alice.public_key)
+    assert b.models.get_outputs_filtered(alice.public_key, spent=False)
