+++ tests/test_txlist.py
@@ -23,7 +23,7 @@ def txlist(b, user_pk, user2_pk, user_sk, user2_sk):
     # Create a TRANSFER transactions
     transfer1 = Transfer.generate(create1.to_inputs(), [([user_pk], 8)], [create1.id]).sign([user2_sk])
 
-    b.store_bulk_transactions([create1, create2, transfer1])
+    b.models.store_bulk_transactions( [create1, create2, transfer1])
 
     return type(
         "",
@@ -37,11 +37,11 @@ def txlist(b, user_pk, user2_pk, user_sk, user2_sk):
 
 @pytest.mark.bdb
 def test_get_txlist_by_asset(b, txlist):
-    res = b.get_transactions_filtered([txlist.create1.id])
+    res = b.models.get_transactions_filtered([txlist.create1.id])
     assert sorted(set(tx.id for tx in res)) == sorted(set([txlist.transfer1.id, txlist.create1.id]))
 
 
 @pytest.mark.bdb
 def test_get_txlist_by_operation(b, txlist):
-    res = b.get_transactions_filtered([txlist.create1.id], operation="CREATE")
+    res = b.models.get_transactions_filtered([txlist.create1.id], operation="CREATE")
     assert set(tx.id for tx in res) == {txlist.create1.id}
